refactor: extract migration dialog to separate component
